Refactor : css파일 정리 및 로그인 버튼 구현
