+++ src/components/css/Card.css
@@ -42,8 +42,6 @@ a {
 }
 
 .text-description {
-  text-overflow: ellipsis;
-  white-space: normal;
   display: -webkit-box;
   -webkit-line-clamp: 3;
   -webkit-box-orient: vertical;
