+++ src/components/css/Main.css
@@ -23,7 +23,6 @@
 
 .search-div {
   color: #666;
-  text-align: right;
   font-family: Pretendard;
   font-size: 16px;
   font-weight: 400;
